tp
Former-commit-id: a98c056b3bc035f3804d7a35f5d3741eb85311ac

+++ src/my_thing_template.h
@@ -58,7 +58,7 @@ public:
         oldptr(this);
     }
 
-    int id {};
+    uint16_t id {};
 
     fsize sz;
